Move ItemTree lowering into its own module
